switch 'float' type from being long double to just double

+++ skull/eval/types/float.c
@@ -11,12 +11,12 @@
 Returns the string representation of float `var`
 */
 char32_t *fmt_float_type(const variable_t *var) {
-	long double data=0;
+	double data=0;
 	variable_read(&data, var);
 
 	if (isinf(data)) {
 		char32_t *ret=NULL;
-		if (data < 0.0L) {
+		if (data < 0.0) {
 			ret=c32sdup(U"-Infinity");
 		}
 		else {
@@ -26,7 +26,7 @@ char32_t *fmt_float_type(const variable_t *var) {
 		return ret;
 	}
 
-	SPRINTF_FMT("%Lf");
+	SPRINTF_FMT("%lf");
 
 	//trim excess zeros off of decimal
 	for (size_t i=strlen(tmp); i>0; i--) {
@@ -49,26 +49,26 @@ char32_t *fmt_float_type(const variable_t *var) {
 Add `lhs` and `rhs` floats together
 */
 variable_t *add_float_type(const variable_t *lhs, const variable_t *rhs) {
-	DO_MATH(U"float", long double, +);
+	DO_MATH(U"float", double, +);
 }
 
 /*
 Subtract `rhs` float from `lhs` float
 */
 variable_t *sub_float_type(const variable_t *lhs, const variable_t *rhs) {
-	DO_MATH(U"float", long double, -);
+	DO_MATH(U"float", double, -);
 }
 
 /*
 Divide `lhs` float by `rhs` float
 */
 variable_t *div_float_type(const variable_t *lhs, const variable_t *rhs) {
-	DO_MATH(U"float", long double, /);
+	DO_MATH(U"float", double, /);
 }
 
 /*
 Multiply `lhs` and `rhs` floats together
 */
 variable_t *mult_float_type(const variable_t *lhs, const variable_t *rhs) {
-	DO_MATH(U"float", long double, *);
+	DO_MATH(U"float", double, *);
 }
